First pass resource_manager refactor
